JL - fix eslint and flow issues
